Fix setting dragFactor BLE value
Fix dragFactor value not being set when no device is subscribed to
notification resulting in empty read value

+++ src/stroke.controller.cpp
@@ -18,12 +18,12 @@ void StrokeController::update()
 {
     if (strokeService.hasDataChanged())
     {
-    cscData = strokeService.getData();
-    if (cscData.revCount != previousRevCount)
-    {
-        Log.infoln("deltaTime: %u", cscData.deltaTime);
-        // Serial.print("deltaTimeDiff: ");
-        // Serial.println(data.deltaTimeDiff);
+        cscData = strokeService.getData();
+        if (cscData.revCount != previousRevCount)
+        {
+            Log.infoln("deltaTime: %u", cscData.deltaTime);
+            // Serial.print("deltaTimeDiff: ");
+            // Serial.println(data.deltaTimeDiff);
         }
     }
 }
@@ -68,7 +68,7 @@ byte StrokeController::getDragFactor() const
     return lround(cscData.dragCoefficient * 1e6);
 }
 
-unsigned long StrokeController::getPreviousRevCount() const
+unsigned int StrokeController::getPreviousRevCount() const
 {
     return previousRevCount;
 }
